+++ src/components/Comment/CommentEdit.jsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import PropTypes from 'prop-types'
 import { Input, Button, ConfigProvider } from 'antd'
 import theme from '@styles/theme'
-import { method } from '@hooks/kyInstance'
+import { method } from '@utils/kyInstance'
 import useKyMutation from '@hooks/useKyMutation'
 
 const inputTheme = {
