+++ src/components/UserProfile/UserFollowButton.jsx
@@ -4,7 +4,7 @@ import { useShallow } from 'zustand/react/shallow'
 import { Button, ConfigProvider, notification } from 'antd'
 import theme from '@styles/theme'
 import useLoginStore from '@stores/useLoginStore'
-import { method } from '@hooks/kyInstance'
+import { method } from '@utils/kyInstance'
 import useKyQuery from '@hooks/useKyQuery'
 import useKyMutation from '@hooks/useKyMutation'
 import useLoginModal from '@hooks/useLoginModal'
