+++ src/hooks/useReviewSubmit.js
@@ -2,7 +2,7 @@ import { useEffect } from 'react'
 import { useNavigate } from 'react-router-dom'
 import uploadToS3 from '@utils/uploadToS3'
 import useFormStore from '@stores/useFormStore'
-import { kyInstance } from './kyInstance'
+import { kyInstance } from '@utils/kyInstance'
 
 const useReviewSubmit = (httpMethod, uri, goPrevStep) => {
   const reviewData = useFormStore((state) => ({
