+++ src/loader/productEditLoader.js
@@ -1,4 +1,4 @@
-import { kyInstance } from '@hooks/kyInstance'
+import { kyInstance } from '@utils/kyInstance'
 
 const productLoader = async ({ params }) => {
   const { id } = params
